requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-generator==1.10
async-timeout==4.0.3
asyncclick==8.1.3.4
attrs==23.1.0
boto3==1.28.62
botocore==1.31.62
certifi==2023.7.22
charset-normalizer==3.3.0
dataclasses-json==0.6.1
Django==4.2.6
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
frozenlist==1.4.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
marshmallow==3.20.1
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
outcome==1.2.0
packaging==23.2
Pillow==10.0.1
pip-review==1.3.0
pluggy==1.3.0
psycopg2==2.9.9
pydantic==1.10.13
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.7.0
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing-extensions==4.8.0
typing-inspect==0.9.0
urllib3==1.26.17
wsproto==1.2.0
yarl==1.9.2
